+++ Factory/SFMLFactory.h
@@ -12,8 +12,8 @@
 class SFMLFactory : public Factory {
 public:
     SFMLFactory(sf::RenderWindow& w, std::shared_ptr<sf::Texture>& t) : window(w), tex(t){};
-    std::shared_ptr<Hiker>  createHiker(std::tuple<float, float> size, std::tuple<float, float> position) override;
-    std::shared_ptr<Entity>  addLane(ColorLogic col, std::tuple<float, float> size, std::tuple<float, float> position) override;
+    std::shared_ptr<Hiker>  createHiker(std::tuple<double, double> size, std::tuple<double, double> position) override;
+    std::shared_ptr<Entity>  addLane(ColorLogic col, std::tuple<double, double> size, std::tuple<double, double> position) override;
 
 private:
     sf::RenderWindow& window;
